requirements.txt
requests
pandas
PyMuPDF
playwright==1.44.0
tqdm
python-dotenv
langchain
duckduckgo-search
langchain-community
pdfplumber
openpyxl>=3.1.2
sentence-transformers>=2.2.2